requirements.txt
fastapi==0.77.1
uvicorn==0.17.6
psycopg2-binary==2.9.3
